Adds lib.size and norm.factors used in fdge back to samples(fdge).
This way we can use the norm.factors and lib.size used in the actual DGE
analysis when visualizing the results downstream from the analaysis.

For example, when we find interesting hits downstream of the DGE analysis,
the data is visualized using the same normalization scheme as was used
in the DGE analysis itself.