initialized(fdgeRunOptions) sensitive to assay types
If an assay_type does not have an fdge method defined, then
`initialized(fdgeRunOptions)` will return FALSE.

I coudn't figure out a way to implement this without introducing state. The
`initialized(fdgeRunOptions)` was not being triggered when the
`updateSelectInput(session, "dge_method")` was resetting the selected thing
to NULL w/ no choices.

Intorducing `state` and reacting to that did the trick.