adds tested_only parameter to samples(ttestmodel) function
Default is FALSE, but when set to TRUE, only the samples used in the contrast
itself are returned.

When the gene ploting (viz and shine) were refactored so that shine relies
on viz, the ttest was showing expression values from all samples used in the
model, and not just the samples used in the contrast.