The fpcaView now coerces dimensions selected to integers.
When you pick axes to plot in the shiny app, I just found that it was returning the character represnetation of the dimension number to plot.

The `pcaviz` reactive would then ultimately call the `viz.FacilePcaAnalysisResult` function with a `dims` parameter that were characters as opposed to integers, and the `assert_integerish` function call now fails.

Not quite sure when/how this happend -- it could be related to the updates to checkmate::check_integerish in ~ Version 2.1.0

https://cran.r-project.org/web/packages/checkmate/news/news.html